backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
Flask-Mail==0.10.0
marshmallow==3.21.3
marshmallow-sqlalchemy==1.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
reportlab==4.2.2
itsdangerous==2.2.0
pytest==8.2.2
pytest-flask==1.3.0
Faker==25.9.1
